Extract issrcML() from a static function to its own file
