+++ src/client/stdin_libarchive.cpp
@@ -8,36 +8,13 @@
  */
 
 #include <stdin_libarchive.hpp>
+#include <issrcml.hpp>
 #include <archive.h>
 #include <SRCMLStatus.hpp>
 #include <libarchive_utilities.hpp>
 #include <string_view>
 
 using namespace ::std::literals::string_view_literals;
-
-namespace {
-    bool issrcML(std::string_view buffer) {
-
-        // short input must be source
-        if (buffer.size() < "<unit"sv.size())
-            return false;
-
-        // trim leading whitespace
-        buffer.remove_prefix(buffer.find_first_not_of(" \n\t"));
-
-        // all XML constructs start with an angle bracket
-        if (buffer[0] != '<')
-            return false;
-
-        // starts with an XML declaration
-        if (buffer[1] == '?')
-            return true;
-
-        // look for start unit
-        return buffer.substr(0, "<src:unit"sv.size()) == "<src:unit"sv ||
-               buffer.substr(0, "<unit"sv.size()) == "<unit"sv;
-    }
-}
 
 void open_stdin_libarchive(srcml_input_src& input_source) {
